requirements.txt
numpy
opencv-python
requests
ultralytics
torch
torchvision
deep_sort_realtime
torchreid
pyzmq
psycopg2-binary
flask
PyYAML==6.0.3
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
